clock: refactor interaction with NTP into a new interface
This refactors HybridClock to interact with the system NTP service (or
its mock) via a new TimeService interface. This encapsulates the
NTP-specific code in such a way that we can easily swap it out with a
built-in NTP client implementation in a future patch.

This also cleans up the 'mock' and OSX (unsynchronized) clock mode to
simply be new implementations of the above interface.

Change-Id: If0653ec32016c588fb63c74f7bc962593aa6c4a0
Reviewed-on: http://gerrit.cloudera.org:8080/7520
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>