consensus: Don't replay config changes
We have invariants in place that make it unnecessary to "replay" a
config change operation, so we will now simply check those invariants
during tablet bootstrap.

This patch removes plumbing of ConsensusMetadataManager into
TabletBootstrap, since it is not needed there, and instead simply passes
in the previously-committed raft config from the consensus metadata in
order to check for correctness.

Change-Id: Id62bf96b21560b2ef5838415e52aeb3720875e62
Reviewed-on: http://gerrit.cloudera.org:8080/7614
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>