Add tablet state summary metrics
This patch adds metrics for the number of tablets in each state
(i.e. the number of tablets for each value of TabletStatePB). To
avoid contention in the tablet manager, the frequency with which
the state numbers are computed is limited. Requests that come
too soon for an update will receive cached counts. The maximum
frequency of updates is controlled by a new flag.

Change-Id: I9fda2061d341586f0aa343787af59984a627080a
Reviewed-on: http://gerrit.cloudera.org:8080/7980
Tested-by: Kudu Jenkins
Reviewed-by: Mike Percy <mpercy@apache.org>