remove 'using std::...' and other from header files
Adding 'using ...' into header files is not a good idea.  In particular,
it confuses the include-what-you-use tool so it gives strange and
incorrect suggestions.  Most likely, there is a bug in the tool as is,
but let's at least not contradict our C++ style guide and remove those
'using ...' from the header files.

This patch does not contain any functional changes.

Change-Id: Iea777af2d881abdb593f9e6f667e613cca8b2fd3
Reviewed-on: http://gerrit.cloudera.org:8080/7628
Reviewed-by: Dan Burkert <danburkert@apache.org>
Tested-by: Alexey Serbin <aserbin@cloudera.com>