[maintenance manager] fix op scheduling lock contention, thread ID tweak
943b1ae26e62a0c82 introduced a lock acquisition on the MM global lock
when a worker thread begins running an op. This resulted in measurable
lock contention with many MM threads, for instance the dense_node-itest
which creates 100 MM threads. This commit introduces a finer grained
lock around the currently running ops container, which reduces the
contention on the global lock.

I've also snuck in a change to make the thread ID correspond to the
system thread ID, instead of a synthetic C++ stdlib ID.

Change-Id: I018ebe65c71344d8911ff66848478f75fef085af
Reviewed-on: http://gerrit.cloudera.org:8080/7621
Reviewed-by: Todd Lipcon <todd@apache.org>
Tested-by: Kudu Jenkins