KUDU-2078: Sink failure if batch size > session's flush buffer size
The Flume sink uses manual flush mode, so if users set the
sink's batch size parameter above the manual flush default
buffer size, the sink could fail batches (over and over). This
patch sets the session's buffer size (which is in terms of
number of ops) to the same as the batch size, so this problem
can no longer occur.

I considered using AUTO_FLUSH_BACKGROUND for the flushing as
well, but it can result in out-of-order writes, which might be
unexpected semantics for Flume (as opposed to, say, Spark).
Using AUTO_FLUSH_BACKGROUND with a high batch size would likely
be more performant, but we can add that as an additional
configuration later if the need arises.

Change-Id: Id1c54bcecc3e13ae64dd90efe6cf53021517dcdf
Reviewed-on: http://gerrit.cloudera.org:8080/7641
Tested-by: Kudu Jenkins
Reviewed-by: Mike Percy <mpercy@apache.org>
Reviewed-on: http://gerrit.cloudera.org:8080/7921
Reviewed-by: Dan Burkert <danburkert@apache.org>