KUDU-2098. Drop Spark 1 Support
Spark 2 has been available for over a year and the value of
shipping a seperate Spark 1 artifact is decreasing.
In the case where another Spark 1 release occurs or bug fixes
are needed, updates to the Kudu Spark 1 integration can still
exist via maintenance releases.

This allows us to clean up deprecated mehods that are used for
compatibility purposes and simplifies the build.

Here is the dev mailing list discussion for reference:
https://s.apache.org/PCfg

Change-Id: I5481cc0477f4d23d89b68ef510a6c9a2aa187537
Reviewed-on: http://gerrit.cloudera.org:8080/7690
Reviewed-by: Dan Burkert <danburkert@apache.org>
Tested-by: Kudu Jenkins