KUDU-2032 (part 2): propagate master hostnames into client
This changes the client code to remember the user-specified master
addresses and propagate them into the creation of master proxies.

It's not possible to reproduce the necessary DNS configurations in a
minicluster test, but with this patch I am now able to use 'kudu perf
loadgen' against a Kerberized cluster even when my local krb5.conf has
rdns=false.

Change-Id: I4bd299f72907a0d9442dd3eb45fcbaa2de4f73ef
Reviewed-on: http://gerrit.cloudera.org:8080/8186
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>