[build-support] IWYU build configuration for Jenkins
In this branch, the IWYU target is just a no-op.  The idea is that
we have common configuration matrix for gerrit pre-commit Jenkins jobs,
so in particular branches the IWYU is run for real (as of now, at least
in the main trunk).

Change-Id: I45f2e8529891dd7da18c9fa6383e0c41e6b4182e
Reviewed-on: http://gerrit.cloudera.org:8080/7750
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins
(cherry picked from commit 8f6812047a2fe50dabfcb26093385b33f9960a7c)
Reviewed-on: http://gerrit.cloudera.org:8080/7778
Reviewed-by: Jean-Daniel Cryans <jdcryans@apache.org>