[build-support] IWYU build configuration for Jenkins
Added provisions to run the include-what-you-use tool against the
updated files in the changelist.  Those are based on the newly
introduced 'iwyu' target to be used like already existing 'lint'
target.

The 'iwyu' target runs IWYU against the compilation database.  So,
as a part of this changelist I updated the top-level CMakeLists.txt
to always generate the compilation database.  The compilation database
is a JSON file containing information on how the targets are built.

Change-Id: I45f2e8529891dd7da18c9fa6383e0c41e6b4182e
Reviewed-on: http://gerrit.cloudera.org:8080/7750
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins