Revert "KUDU-2065, KUDU-2011: Release sidecars on cancellation or timeout"
This reverts commit 351337ee2d92361e2c87699d1c87a49774a3ab4d.

This patch seemed to cause a problem in test cluster that runs YCSB
stress tests. An RPC client and server got "out of sync" somehow on a
connection so that the footer of one call was interpreted as the length
prefix of the next call by the server. It then expected a 40MB+ RPC and
just sat there consuming RPC calls for many hours assuming it was all
part of one large RPC. The other end just kept timing out since it was
sending calls and not getting any responses.

Michael and I looked into it for a while and found one issue but still
couldn't fully explain how this happened, so we'll revert for now.

Change-Id: I1430cabfc54e858c1ad073574404821ae71c169f
Reviewed-on: http://gerrit.cloudera.org:8080/7788
Tested-by: Todd Lipcon <todd@apache.org>
Reviewed-by: Todd Lipcon <todd@apache.org>