Bump test timeout for flex_partitioning-itest
In precommit runs, this test is sharded 8 ways so it never has issues with
timeouts. However, in some Jenkins builds that don't use dist-test,
it can time out, especially when TSAN is enabled.

This adds functionality to our ADD_KUDU_TEST() cmake function to specify
a timeout and plumbs that timeout through to our test wrapper shell
script.

Tested by verifying that the appropriate environment variable gets set
in the environment of the running test.

Change-Id: I07661415d3500dfa5d7984b720b7a6f14597cc19
Reviewed-on: http://gerrit.cloudera.org:8080/7814
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Todd Lipcon <todd@apache.org>