[build-support] fixed IWYU no-op configuration
Jenkins is upset if no output is produced by a build: no artifacts
to archive, so it fails the build in that case.

This patch adds an informational no-op message to give Jenkins some toys
to play and make it happy.

This is a follow-up for 4bf1225e8a95e10e341a9e30b29b8ae0e36d60ce

Change-Id: Id349a1ee83ad6b9e5fb6f35c4d044487e6aff1c3
Reviewed-on: http://gerrit.cloudera.org:8080/7833
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>