Fix flakiness in ts_recovery-itest TestCrashBeforeWriteLogSegmentHeader
TestCrashBeforeWriteLogSegmentHeader injects server faults during a
write workload before writing the log segment header. In a stressed test
environment, writes may take longer, and workloads may time out before
even communicating with the server. If the workload never successfully
writes to the server for this reason, and never makes it to the
fault-injection-site, the test will fail.

This is resolved by bumping the timeout allowed by the workload,
ensuring there is enough time to get the table locations and insert data
to them.

With the bumped timeout, this passes 500/500 runs in TSAN with
--stress_cpu_threads=8, rather than failing ~75%.

Change-Id: I382bb1109ef12c23d6f191b64d97bcaa6ec96d6a
Reviewed-on: http://gerrit.cloudera.org:8080/7838
Reviewed-by: Todd Lipcon <todd@apache.org>
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Kudu Jenkins