[build-support] fixed IWYU no-op configuration
Jenkins is upset if no output is produced by a build: no artifacts
to archive, so it fails the build in that case.

This patch adds an informational no-op message to give Jenkins some toys
to play and make it happy.

This is a follow-up for 640e89767faabdec04cc1486f87af2814f12986d

Change-Id: Iafa8f3c0d024fe650344a7b956705f3684dd1116
Reviewed-on: http://gerrit.cloudera.org:8080/7832
Tested-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Todd Lipcon <todd@apache.org>