[build-support] fixed IWYU no-op configuration
Jenkins is upset if no output is produced by a build: no artifacts
to archive, so it fails the build in that case.

This patch adds an informational no-op message to give Jenkins some toys
to play and make it happy.

This is a follow-up for 8e3a558d1661fb9b985143bb7a8d54fd0250b03b

Change-Id: I0f7dff27981f70b7247f44d7e6bfe0c99a78548e
Reviewed-on: http://gerrit.cloudera.org:8080/7831
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>