KUDU-2102: fix PosixRWFile::Sync to guarantee durability when used concurrently
PosixRWFile has an AtomicBool to optimize the perfomance of Sync()
so that when two threads call Sync() at the same time, only one will
actually do the fsync(). However, the optimization may result in one
thread returning from the call early and operating under the assumption
that the file's data has been made durable, even though it is still in
the process of being synchronized to disk.

This patch fixes this issue simply by removing the optimization. Since
the overhead of no-op fsync() is purely syscall-related, this optimization
wasn't meaningful to begin with.

Change-Id: Ifaf2686233b2553de6d4a4c76fa67cc48dd340d4
Reviewed-on: http://gerrit.cloudera.org:8080/7835
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Adar Dembo <adar@cloudera.com>