KUDU-1544: Race in Java client's AsyncKuduSession.apply()
This fixes a potential race in AsyncKuduSession.apply by acquiring the
notification deferred early. See the JIRA for details about the race.
Acquiring the notification deferred early is always safe, the only
downside is that clients may be spuriously notified that there is buffer
space available when in fact there is not.

Change-Id: I1791f5ed51e2d79cd9e9bcbf83e9b30eede871b4
Reviewed-on: http://gerrit.cloudera.org:8080/7839
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins